Try fixing Ubuntu not finding FFTW libraries (2).
