+++ SharpFFTW.Tests/Setup.cs
@@ -1,7 +1,4 @@
 ﻿using NUnit.Framework;
-using System.Reflection;
-using System.Runtime.InteropServices;
-using System;
 
 namespace SharpFFTW.Tests
 {
@@ -11,19 +8,7 @@ namespace SharpFFTW.Tests
         [OneTimeSetUp]
         public void GlobalSetup()
         {
-            NativeLibrary.SetDllImportResolver(Assembly.GetExecutingAssembly(), DllImportResolver);
-        }
-
-        private static IntPtr DllImportResolver(string library, Assembly assembly, DllImportSearchPath? searchPath)
-        {
-            // Trying to find FFTW on Ubuntu.
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux) && library.StartsWith("fftw"))
-            {
-                return NativeLibrary.Load("lib" + library + ".so.3", assembly, searchPath);
-            }
-
-            // Otherwise, fall back to default import resolver.
-            return IntPtr.Zero;
+            Library.SetImportResolver();
         }
     }
 }
